Added IRdq to DAIM
Added IRdq to DAIMIB, or rather DAIMIB2 calls DAIM_IRdq for the ODEs and DAE terms. DAIM_IRdq specifically has a modified current_control_extref function that takes vg_d and vg_q inputs (y values) instead of the vg times sin or cos as in the D'Arco code.

I did possibly 'revert' some DAIM and DArcoEPSR122 code because I think was following an older plan of generalization, and I just wanted it to work as a standalone for comparison.

Still might not be working perfectly e.g. checking vg or vo might not be correct.. looking to explore more later.